requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
google-cloud-discoveryengine==0.11.11
python-dotenv==1.0.0
pydantic==2.5.0
aiohttp==3.9.1
asyncio==3.4.3
jinja2==3.1.2
python-multipart==0.0.6 